py-hr-connect-bot 2/requirements.txt
Flask
requests
python-dateutil
pymongo
pytz
cryptography
flask-cors
gpt-index
pyOpenSSL
transformers
pycryptodomex
pypdf
orjson
pymongo
